+++ photoViewer/package.json
@@ -11,18 +11,18 @@
     "test": "jest"
   },
   "dependencies": {
-    "react": "^17.0.0",
-    "react-dom": "^17.0.0"
+    "react": "^19.0.0",
+    "react-dom": "^19.0.0"
   },
   "devDependencies": {
     "@eslint/js": "^9.19.0",
     "@testing-library/jest-dom": "^6.6.3",
-    "@testing-library/react": "^12.1.0",
+    "@testing-library/react": "^16.2.0",
     "@testing-library/user-event": "^14.0.0",
     "@types/jest": "^29.5.14",
-    "@types/react": "^17.0.8",
-    "@types/react-dom": "^17.0.3",
-    "@types/react-test-renderer": "^17.0.0",
+    "@types/react": "^19.0.8",
+    "@types/react-dom": "^19.0.3",
+    "@types/react-test-renderer": "^19.0.0",
     "@vitejs/plugin-react": "^4.3.4",
     "eslint": "^9.19.0",
     "eslint-plugin-react-hooks": "^5.0.0",
@@ -31,7 +31,7 @@
     "identity-obj-proxy": "^3.0.0",
     "jest": "^29.7.0",
     "jest-environment-jsdom": "^29.7.0",
-    "react-test-renderer": "^17.0.0",
+    "react-test-renderer": "^19.0.0",
     "ts-jest": "^29.2.5",
     "typescript": "~5.7.2",
     "typescript-eslint": "^8.22.0",
