+++ photoViewer/src/image-thumbnails/imageThumbnails.css
