+++ photoViewer/src/large-photo-viewer/mainPhoto.css
